move functions in files.ts to writer.ts and reader.ts
