+++ lib/controller.ts
@@ -11,7 +11,8 @@ import Cache from './storage/caching/cache';
 import { error, fileForPage } from './utils/utils';
 import rdfSerializer from 'rdf-serialize';
 import rdfParser from 'rdf-parse';
-import { convert, convertToJsonLD } from './storage/file-system/files';
+import { convert } from './storage/file-system/reader';
+import { convertToJsonLD } from './storage/file-system/reader';
 import PromiseQueue from './utils/promise-queue';
 import Node from './models/node';
 import convertToMember from './converters/member-converter';
